+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/TornadoTargetDevice.java
@@ -55,7 +55,7 @@ public interface TornadoTargetDevice {
 
     long[] getDeviceMaxWorkGroupSize();
 
-    long getMaxThreadsPerBlock();
+    int getMaxThreadsPerBlock();
 
     int getDeviceMaxClockFrequency();
